Addressed suggestion made by Zilong Chang in StJetMaker2012.cxx, StjFCSMuDst.h/cxx. Have yet to add mstu delimeters from StarGenEvent in StjMCMuDst.cxx.

+++ StRoot/StJetMaker/mudst/StjFCSMuDst.h
@@ -18,7 +18,7 @@ public:
     StjFCSMuDst();
     virtual ~StjFCSMuDst() {}
 
-    void setVertex(float vx, float vy, float vz)
+    void setVertex(double vx, double vy, double vz)
     {
       _setVertex = true;
       _vx = vx;
